+++ .gitignore
@@ -1,1 +1,6 @@
 node_modules
+
+lib
+dist
+
+*.log
