+++ package.json
@@ -2,10 +2,15 @@
   "name": "mafmt",
   "version": "1.0.1",
   "description": "A multiaddr validator",
-  "main": "src/index.js",
+  "main": "lib/index.js",
+  "jsnext:main": "src/index.js",
   "scripts": {
-    "lint": "standard",
-    "test": "node tests/test.js"
+    "lint": "dignified-lint",
+    "build": "dignified-build",
+    "test": "dignified-test",
+    "test:node": "dignified-test node",
+    "test:browser": "dignified-test browser",
+    "release": "dignified-release"
   },
   "pre-commit": [
     "lint",
@@ -25,9 +30,9 @@
   },
   "homepage": "https://github.com/whyrusleeping/js-mafmt#readme",
   "devDependencies": {
-    "pre-commit": "^1.1.2",
-    "standard": "^6.0.8",
-    "tape": "^4.4.0"
+    "chai": "^3.5.0",
+    "dignified.js": "github:dignifiedquire/dignified.js",
+    "pre-commit": "^1.1.2"
   },
   "dependencies": {
     "multiaddr": "^1.3.0"
